requirements.txt
fastapi
uvicorn
python-dotenv
paramiko
zipfile36
rarfile